+++ tests/inter_channel_cli.rs
@@ -71,7 +71,7 @@ fn cli_keys(slot: u8) -> MemberKeys {
 
 #[derive(Serialize, Deserialize, Clone)]
 struct ControlledMember {
-    slot: u8,
+    slot: u16,
     keygen_seed: u64,
     balance_amount: u64,
     balance_seed: u64,
@@ -88,7 +88,7 @@ struct CliState {
     spent_tx_hashes: Vec<Bytes32>,
 }
 
-fn member_info(slot: u8, keys: &MemberKeys) -> MemberInfo {
+fn member_info(slot: u16, keys: &MemberKeys) -> MemberInfo {
     MemberInfo {
         slot,
         pk_g: keys.pk_g(),
@@ -117,7 +117,7 @@ fn build_cli_channel(channel_id: u32, balances: &[u64]) -> ChannelFixture {
     let members: Vec<MemberInfo> = keys
         .iter()
         .enumerate()
-        .map(|(i, k)| member_info(i as u8, k))
+        .map(|(i, k)| member_info(i as u16, k))
         .collect();
     let record = build_record(channel_id, &members, 0, 0).expect("record");
 
@@ -135,7 +135,7 @@ fn build_cli_channel(channel_id: u32, balances: &[u64]) -> ChannelFixture {
         cts.push(ct);
         witnesses.push(w);
         controlled.push(ControlledMember {
-            slot: i as u8,
+            slot: i as u16,
             keygen_seed: cli_keygen_seed(i as u8),
             balance_amount: bal,
             balance_seed,
@@ -229,8 +229,8 @@ fn build_transfer(
     a_balances: &[u64],
     a_witnesses: &[intmax3_zkp::regev::AmountWitness],
     b_record: &ChannelRecord,
-    sender_slot: u8,
-    recipient_slot: u8,
+    sender_slot: u16,
+    recipient_slot: u16,
     amount: u64,
     nullifier_tag: u32,
     rng_seed: u64,
@@ -275,7 +275,7 @@ fn build_burn_send_debits_only_sender_and_targets_l1() {
 
     let a = build_cli_channel(A_ID, &[50, 10, 30]);
     let a_keys: Vec<MemberKeys> = CLI_SLOTS.iter().map(|&s| cli_keys(s)).collect();
-    let sender_slot = 0u8;
+    let sender_slot = 0u16;
     let amount = 20u64;
     let l1 = Address::from_hex("0x00000000000000000000000000000000000000aa").unwrap();
     let prev_fund = a.snapshot.state.channel_fund.amount;
@@ -344,7 +344,7 @@ fn burn_send_base_chain_matches_channel() {
 
     let a = build_cli_channel(A_ID, &[50, 10, 30]);
     let a_keys: Vec<MemberKeys> = CLI_SLOTS.iter().map(|&s| cli_keys(s)).collect();
-    let sender_slot = 0u8;
+    let sender_slot = 0u16;
     let amount = 20u64;
     let l1 = intmax3_zkp::ethereum_types::address::Address::from_hex(
         "0x00000000000000000000000000000000000000aa",
@@ -417,8 +417,8 @@ fn inter_channel_cli_end_to_end() {
     // ---- Build both channels (members = the binary's CLI members) + seed cli_state.json. ----
     let a = build_cli_channel(A_ID, &[50, 10, 30]);
     let b = build_cli_channel(B_ID, &[20, 40, 60]);
-    let sender_slot = 0u8;
-    let recipient_slot = 1u8;
+    let sender_slot = 0u16;
+    let recipient_slot = 1u16;
 
     let b_keys: Vec<MemberKeys> = CLI_SLOTS.iter().map(|&s| cli_keys(s)).collect();
     let b_record = b.record.clone();
@@ -566,8 +566,8 @@ fn inter_channel_cli_forged_a_state_refused() {
 
     let a = build_cli_channel(A_ID, &[50, 10, 30]);
     let b = build_cli_channel(B_ID, &[20, 40, 60]);
-    let sender_slot = 0u8;
-    let recipient_slot = 1u8;
+    let sender_slot = 0u16;
+    let recipient_slot = 1u16;
 
     let b_keys: Vec<MemberKeys> = CLI_SLOTS.iter().map(|&s| cli_keys(s)).collect();
     let b_record = b.record.clone();
